# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User Registration API",
    "version": "1.0.0",
    "description": "API for ACL Project"
  },
  "servers": [
    {
      "url": "http://localhost:3030"
    }
  ],
  "paths": {
...
    "/addProduct/{userId}": {
      "post": {
        "summary": "Add a new product",
        "description": "Adds a new product by a user with the specified userId.",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "The ID of the user adding the product.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "picture": {
                    "type": "string",
                    "description": "The URL of the product picture."
                  },
                  "price": {
                    "type": "number",
                    "format": "float",
                    "description": "The price of the product."
                  },
                  "description": {
                    "type": "string",
                    "description": "A description of the product."
                  },
                  "sellerType": {
                    "type": "string",
                    "description": "The type of the seller."
                  },
                  "originalQuantity": {
                    "type": "integer",
                    "description": "The original quantity of the product available."
                  },
                  "ratings": {
                    "type": "number",
                    "description": "The initial ratings of the product."
                  },
                  "reviews": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "List of reviews for the product."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Product added successfully"
                    },
                    "product": {
                      "type": "object",
                      "description": "The created product object.",
                      "properties": {
                        "picture": {
                          "type": "string",
                          "description": "The URL of the product picture."
                        },
                        "price": {
                          "type": "number",
                          "format": "float",
                          "description": "The price of the product."
                        },
                        "description": {
                          "type": "string",
                          "description": "A description of the product."
                        },
                        "sellerType": {
                          "type": "string",
                          "description": "The type of the seller."
                        },
                        "originalQuantity": {
                          "type": "integer",
                          "description": "The original quantity of the product available."
                        },
                        "ratings": {
                          "type": "number",
                          "description": "The initial ratings of the product."
                        },
                        "reviews": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "List of reviews for the product."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields or invalid user type.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Price, picture, original quantity, description, sellerType, name, ratings and reviews are required."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error adding product"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/getAvailableProducts/{userId}": {
      "get": {
        "summary": "Get available products",
        "description": "Retrieves all available products for a user with the specified userId.",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "The ID of the user retrieving available products.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of available products.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "picture": {
                        "type": "string",
                        "description": "The URL of the product picture."
                      },
                      "price": {
                        "type": "number",
                        "format": "float",
                        "description": "The price of the product."
                      },
                      "description": {
                        "type": "string",
                        "description": "A description of the product."
                      },
                      "sellerType": {
                        "type": "string",
                        "description": "The type of the seller."
                      },
                      "ratings": {
                        "type": "number",
                        "description": "The product's ratings."
                      },
                      "reviews": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "description": "List of reviews for the product."
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid user type.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid type"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error retrieving available products"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/filterProducts/{userId}": {
      "get": {
        "summary": "Filter products by price range",
        "description": "Retrieves products within a specified price range for a user with the specified userId.",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "The ID of the user filtering products.",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minPrice",
            "required": false,
            "description": "The minimum price of products to retrieve.",
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "in": "query",
            "name": "maxPrice",
            "required": false,
            "description": "The maximum price of products to retrieve.",
            "schema": {
              "type": "number",
              "format": "float"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products filtered by price range.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "picture": {
                        "type": "string",
                        "description": "The URL of the product picture."
                      },
                      "price": {
                        "type": "number",
                        "format": "float",
                        "description": "The price of the product."
                      },
                      "description": {
                        "type": "string",
                        "description": "A description of the product."
                      },
                      "sellerType": {
                        "type": "string",
                        "description": "The type of the seller."
                      },
                      "ratings": {
                        "type": "number",
                        "description": "The product's ratings."
                      },
                      "reviews": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "description": "List of reviews for the product."
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid user type.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid type"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error querying products by price range"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/searchProductByName/{userId}": {
      "get": {
        "summary": "Search products by name",
        "description": "Searches for products by name for a user with the specified userId.",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "The ID of the user searching for products.",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": true,
            "description": "The name of the product to search for.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products matching the search criteria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "picture": {
                        "type": "string",
                        "description": "The URL of the product picture."
                      },
                      "price": {
                        "type": "number",
                        "format": "float",
                        "description": "The price of the product."
                      },
                      "description": {
                        "type": "string",
                        "description": "A description of the product."
                      },
                      "sellerType": {
                        "type": "string",
                        "description": "The type of the seller."
                      },
                      "ratings": {
                        "type": "number",
                        "description": "The product's ratings."
                      },
                      "reviews": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "description": "List of reviews for the product."
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid user type or missing product name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid type or missing product name"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error searching for products"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}